refactor status tags into components for each viewer type
