+++ app/components/appropriate_bodies/participants/table_row.html.erb
@@ -30,6 +30,6 @@
     <%= training_status %>
   </td>
   <td class="govuk-table__cell">
-    <%= status_tag %>
+    <%= render StatusTags::AppropriateBodyParticipantStatusTag.new(participant_profile:, induction_record:) %>
   </td>
 </tr>
