+++ app/components/status_tags/school_participant_status_tag.html.erb
