+++ app/views/admin/schools/participants/index.html.erb
@@ -22,7 +22,7 @@
           <td class="govuk-table__cell"><%= t participant.participant_type, scope: "schools.participants.type" %></td>
           <td class="govuk-table__cell"><%= participant.cohort&.start_year %></td>
           <td class="govuk-table__cell">
-            <%= render ParticipantStatusTagComponent.new(profile: participant, induction_record: participant.induction_records.for_school(@school).latest) %>
+            <%= render StatusTags::ParticipantStatusTag.new(profile: participant, induction_record: participant.induction_records.for_school(@school).latest) %>
           </td>
         </tr>
       <% end %>
